refactor: modulariza arquivos de teste
